requirements.txt
fastapi
uvicorn
pydicom
numpy
Pillow
python-multipart
matplotlib
scipy
docopt
requests